GET: Retrieve details of a specific match by its ID

+++ src/main/java/andrei/springboot/fotbal_proiect/dao/MatchRepository.java
@@ -3,5 +3,5 @@ package andrei.springboot.fotbal_proiect.dao;
 import andrei.springboot.fotbal_proiect.entity.Match;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MatchRepository extends JpaRepository <Match, Integer> {
+public interface MatchRepository extends JpaRepository <Match, Long> {
 }
